requirements.txt
# Django
Django==4.2.6
djangorestframework==3.14.0
drf-yasg==1.21.7
asgiref==3.7.2
typing_extensions==4.8.0
sqlparse==0.4.4

# Red neuronal
numpy==1.26.0
tensorflow==2.14.0
tensorflow-text==2.14.0
keras==2.14.0
scikit-learn==1.3.1
pandas==2.0.3

# Other
gunicorn==21.2.0
matplotlib==3.8.0
Faker==18.11.2
pydot==1.4.2